requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ML
torch==2.1.2
numpy==1.26.3

# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0

# HTTP Client
requests==2.31.0

# Utils
python-dotenv==1.0.0
